AQMS_Tx
This is the transmitter-side code for an IoT-based Air Quality Monitoring System (AQMS). It reads real-time data from MQ135 and MQ2 gas sensors using an ESP32 and transmits it over LoRa communication to a receiver node. The setup is optimized for low power consumption, making it ideal for solar-powered outdoor deployments. Designed for long-range, stable, and energy-efficient data transmission in air quality monitoring networks.