AQMS_Rx
This is the receiver-side code for an IoT-based Air Quality Monitoring System (AQMS). It uses an ESP32 with LoRa module to receive air quality data transmitted from remote sensor nodes. The received data is processed and sent to the ThingSpeak IoT platform for real-time monitoring and visualization. Designed for reliable reception, cloud integration, and seamless operation in distributed AQMS networks.